// apps/frontend/src/api/generated/model/generateCommunicationRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Alittlebyte PoC API
 * Versioned JSON contract consumed by the React SPA.
 * OpenAPI spec version: 1.0.0
 */
import type { GenerateCommunicationRequestStyle } from './generateCommunicationRequestStyle';
import type { GenerateCommunicationRequestTone } from './generateCommunicationRequestTone';

export interface GenerateCommunicationRequest {
  /**
     * @minLength 12
     * @maxLength 5000
     */
  prompt: string;
  tone: GenerateCommunicationRequestTone;
  style: GenerateCommunicationRequestStyle;
}